fix: split hook types and utils.
