+++ src/index.ts
@@ -2,3 +2,5 @@ import useValidations from './useValidations';
 
 export default useValidations;
 export * from './useValidations';
+export * from './types';
+export * from './validators';
