+++ src/validators.ts
@@ -1,0 +1,3 @@
+import { isEmpty } from "./utils";
+
+export const isRequired = (value: string) => (isEmpty(value) ? "required" : null);
